例外処理を追加、メモを削除で記号、空白の入力拒否の処理を追加

+++ app.rb
@@ -19,14 +19,6 @@ pattern = /[%.+\/\p{Space}]/
 #string型で標準入力
 input = Readline.readline("指示を「 1 」か「 2 」か「3」か「 4 」で入力してください:").to_s
 
-#input = Integer.sqrt(input)
-
-=begin
-normalized_input = true if input.match?(pattern) == true
-normalized_input = input if input.match?(pattern) == false
-=end
-
-
 def normalized(input, pattern)
   if input.match?(pattern) == true
     return true
@@ -114,8 +106,12 @@ end
 #メモを削除する処理
 when input == "2"
   puts "メモのタイトル一覧"
+
   #メモの保存ディレクトリを配列化
   file_dir = Dir.entries(dir_path)
+  
+  #不要な要素を取り除いて、配列を再生成
+  file_dir = file_dir.drop(2)
   
    #メモの保存ディレクトリを標準出力
   file_dir.each do |file| 
@@ -126,11 +122,27 @@ when input == "2"
   loop do
   #メモのタイトルを標準入力で受け取る
   file_name = Readline.readline("削除したいファイル名を入力してください。＞ :")
-  #file_dirの配列から標準入力されたタイトルを比較、真偽値を格納
-  file_status = file_dir.include?(file_name)
-  
-  #真偽値で処理を分岐、tureなら標準入力で受け取ったタイトルのメモの内容を表示する、falseならメモを見るの処理に戻る
-  if file_status == true
+
+#file_nameを正規表現でマッチング、変数に真偽値を格納
+normalized_input = file_name.match?(pattern)
+    
+#file_dirの配列から標準入力されたタイトルを比較、変数に真偽値を格納
+value_exists = file_dir.include?(file_name)
+
+#関数処理、引数の型によって別て戻り値に真偽値を返す
+def normalized(value_exists, normalized_input)
+if value_exists == true
+  return true
+elsif normalized_input == false
+  return false
+end
+end
+
+#関数の戻り値が変数に代入
+normalized_input = normalized(value_exists, normalized_input)
+
+#真偽値で処理を分岐、tureなら標準入力で受け取ったタイトルのメモの内容を表示する、falseならメモを見るの処理に戻る
+if normalized_input == true
   file_path = File.join(dir_path, file_name)
   File.delete(file_path)
   puts "#{file_name}を削除しました。" if File.delete == 0
@@ -139,17 +151,17 @@ when input == "2"
   end
   
   #file_status配列とfile_nameを比較して戻り値がtureならメモを見るの処理の繰り返しを終了
-  break if file_status == true
+  break if normalized_input == true
   #メモを表示するのループ終了
-  
 end
+
 #メモの内容を表示する処理
 when input == "3"
   puts "メモのタイトル一覧"
   #メモの保存ディレクトリを配列化
   file_dir = Dir.entries(dir_path)
   
-  #
+  #不要な要素を取り除いて、配列を再生成
   file_dir = file_dir.drop(2)
   
   #メモの保存ディレクトリを標準出力
